+++ lib/src/languages/java/java8/java8_generator.dart
@@ -139,15 +139,83 @@ class ApolloCodeGeneratorJava8 extends ApolloCodeGenerator {
     s.write(indent);
 
     var str = value.value;
+    str = _escapeString(str);
+    s.write('"$str"');
 
-    str = str
+    return s;
+  }
+
+  String _escapeString(String str) {
+    return str
         .replaceAll('\t', r'\t')
         .replaceAll('"', r'\"')
         .replaceAll('\r', r'\r')
         .replaceAll('\n', r'\n')
         .replaceAll('\b', r'\b');
+  }
 
-    s.write('"$str"');
+  @override
+  StringBuffer generateASTValueStringConcatenation(
+      ASTValueStringConcatenation value,
+      [String indent = '',
+      StringBuffer? s]) {
+    var list = <dynamic>[];
+
+    for (var v in value.values) {
+      if (v is ASTValueStringVariable) {
+        var s2 = generateASTValueStringVariable(v, '');
+        list.add(s2);
+      } else if (v is ASTValueStringExpresion) {
+        var s2 = generateASTValueStringExpresion(v, '');
+        list.add(s2);
+      } else if (v is ASTValueStringConcatenation) {
+        var s2 = generateASTValueStringConcatenation(v, '');
+        list.add(s2.toString());
+      } else if (v is ASTValueString) {
+        var s2 = generateASTValueString(v, '');
+        list.add(s2.toString());
+      }
+    }
+
+    s ??= StringBuffer();
+
+    for (var i = 1; i < list.length;) {
+      var prev = list[i - 1];
+      var e = list[i];
+
+      if (prev is String && e is String) {
+        var merge = prev.substring(0, prev.length - 1) + e.substring(1);
+        list[i - 1] = merge;
+        list.removeAt(i);
+      } else {
+        ++i;
+      }
+    }
+
+    for (var i = 0; i < list.length; ++i) {
+      var e = list[i];
+      if (i > 0) s.write(' + ');
+      s.write(e);
+    }
+
+    return s;
+  }
+
+  @override
+  StringBuffer generateASTValueStringVariable(ASTValueStringVariable value,
+      [String indent = '', StringBuffer? s]) {
+    s ??= StringBuffer();
+    s.write('String.valueOf( ${value.variable.name} )');
+    return s;
+  }
+
+  @override
+  StringBuffer generateASTValueStringExpresion(ASTValueStringExpresion value,
+      [String indent = '', StringBuffer? s]) {
+    s ??= StringBuffer();
+
+    var exp = generateASTExpression(value.expression, '').toString();
+    s.write('String.valueOf( $exp )');
 
     return s;
   }